+++ packages/read-orders/backend/node-express/server.js
@@ -2,6 +2,7 @@ const express = require('express');
 const cors = require('cors');
 const dotenv = require('dotenv');
 const mockDb = require('./utils/mock-db');
+const { isOrderEmail, prepEmailForParser } = require('./utils/email-helpers');
 const axios = require('axios');
 
 const Nylas = require('nylas');
@@ -55,25 +56,6 @@ const expressBinding = new ServerBindings.express(nylasClient, {
   },
   clientUri: CLIENT_URI,
 });
-
-function isOrderEmail(message) {
-  return ['order', 'purchase', 'tracking'].some((word) =>
-    message.subject.toLowerCase().includes(word)
-  );
-}
-
-function prepEmailForParser(messageObj, rawEmail) {
-  const encodeBase64 = (data) => Buffer.from(data).toString('base64');
-  const senderEmail = messageObj.from[0].email;
-
-  return {
-    emailTimestamp: +messageObj.date * 1000,
-    fetchedEmailId: messageObj.id,
-    from: senderEmail,
-    senderDomain: senderEmail.split('@')[1],
-    textBase64: encodeBase64(rawEmail),
-  };
-}
 
 // Mount the express middleware to your express app
 const nylasMiddleware = expressBinding.buildMiddleware();
